fix conversion in mmol/L for y axis labels in live activity charts
improve readability for lockscreen widget in light mode

+++ FreeAPS/Sources/Services/LiveActivity/LiveActitiyShared.swift
@@ -7,7 +7,7 @@ struct LiveActivityAttributes: ActivityAttributes {
         let trendSystemImage: String?
         let change: String
         let date: Date
-        let chart: [Int16]
+        let chart: [Double]
         let chartDate: [Date?]
         let rotationDegrees: Double
     }
